Implement local website development and AWS Lambda function
Local website development and AWS Lambda function products
both rely on an internal dependency called `GenerateWebsite`,
which is the common code to generate HTML when called.

Note: the AWS Lambda function product (Swift-AWS-Lambda-Website)
needs to be built against Linux to work on AWS Lambda. This will
be done using Docker!